Avoid some more direct pointer manipulation

+++ common/alcomplex.cpp
@@ -206,7 +206,7 @@ void complex_hilbert(const al::span<std::complex<double>> buffer)
 
     const double inverse_size = 1.0/static_cast<double>(buffer.size());
     auto bufiter = buffer.begin();
-    const auto halfiter = bufiter + (buffer.size()>>1);
+    const auto halfiter = bufiter + ptrdiff_t(buffer.size()>>1);
 
     *bufiter *= inverse_size; ++bufiter;
     bufiter = std::transform(bufiter, halfiter, bufiter,
